login

+++ myweb/base/views.py
@@ -1,13 +1,15 @@
 from django.shortcuts import render, redirect
 from django.http import HttpResponse
 # Create your views here.
-from .models import Room, Topic
+from .models import Room, Topic, Message
 from .forms import RoomForm
 from django.db.models import Q
 from django.contrib.auth.models import User
 from django.contrib import messages
 from django.contrib.auth import authenticate, login, logout
-
+from django.contrib.auth.decorators import login_required
+from django.http import HttpResponse
+from django.contrib.auth.forms import UserCreationForm
 
 # rooms = [
 #     {'id': 1, "name": "Let's learn Python!"},
@@ -46,10 +48,22 @@ def home(request):
 def room(request, pk):
     # return HttpResponse("<h1>Room<h1>")
     room = Room.objects.get(id=int(pk))
-    context = {"room": room}
+    room_messages = room.message_set.all().order_by('-created') #get messages which are related to room
+
+    #message creation
+    if request.method == "POST":
+        message = Message.objects.create(
+            user=request.user,
+            room=room,
+            body=request.POST.get('body')
+        )
+        return redirect('room', pk=room.id)
+
+    context = {"room": room, 'room_messages': room_messages}
     return render(request, "base/room.html", context)
 
 
+@login_required(login_url='login')
 def create_room(request):
     form = RoomForm()
     if request.method == "POST":
@@ -62,10 +76,14 @@ def create_room(request):
     context = {'form': form}
     return render(request, 'base/room_form.html', context)
 
-
+@login_required(login_url='login')
 def update_room(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+
+    if request.user != room.host:
+        return HttpResponse("<h1>You Don't Have Permission!</h1>")
+
     if request.method == "POST":
         form = RoomForm(request.POST, instance=room)
         if form.is_valid():
@@ -74,9 +92,13 @@ def update_room(request, pk):
     context = {'form': form}
     return render(request, 'base/room_form.html', context)
 
-
+@login_required(login_url='login')
 def delete_room(request, pk):
     room = Room.objects.get(id=pk)
+
+    if request.user != room.host:
+        return HttpResponse("<h1>You Don't Have Permission!</h1>")
+
     if request.method == "POST":
         room.delete()
         return redirect('home')
@@ -84,8 +106,13 @@ def delete_room(request, pk):
 
 
 def login_page(request):
+    #ბოლოს
+    page = "login"
+    if request.user.is_authenticated:
+        return redirect('home')
+
     if request.method == "POST":
-        username = request.POST.get('username')
+        username = request.POST.get('username').lower()
         password = request.POST.get('password')
         try:
             user = User.objects.get(username=username)
@@ -98,9 +125,24 @@ def login_page(request):
         else:
             messages.error(request, "Username or Password doesn't exist!")
 
-    context = {}
+    context = {"page": page}
     return render(request, "base/login_register.html", context)
 
 def logout_user(request):
     logout(request)
     return redirect('home')
+
+def register_page(request):
+    form = UserCreationForm()
+    if request.method == "POST":
+        form = UserCreationForm(request.POST)
+        if form.is_valid():
+            user = form.save(commit=False)
+            user.username = user.username.lower()
+            user.save()
+            login(request, user)
+            return redirect('home')
+        else:
+            messages.error(request, "An error occurred during registration")
+
+    return render(request, 'base/login_register.html', {'form': form})
